requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.32.5
openai==0.28.1
line-bot-sdk==3.5.0